pvp: refactor attack data
